pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "recursivedisgustingsystemcall"
version = "0.1.0"
[tool.poetry.dependencies]
python = "^3.8"
request = "^2019.4"
similar-text = "^0.2.0"
[tool.poetry.dev-dependencies]
